requirements.txt
streamlit==1.39.0
pillow==10.4.0
openai==1.52.0
beautifulsoup4==4.12.3
gspread==6.1.2
oauth2client==4.1.3
pandas==2.2.3
xlsxwriter==3.2.0
requests==2.32.3
python-dotenv==1.0.1
